# OpenAPI/saiboard-admin/openapi.yaml
{
    "host": "",
    "basePath": "",
    "schemes": [
        "http"
    ],
    "openapi": "3.0.1",
    "info": {
        "title": "B8AIadmin 接口文档",
        "description": "基于 SaiAdmin/Webman 控制器注解生成，供接口调试与 uniapp 自动接口生成使用。",
        "version": "1.0.7"
    },
...
    "paths": {
        "/app/saiboard/admin/Screen/index": {
            "get": {
                "tags": [
                    "saiboard-admin.SAI Board.ScreenController"
                ],
                "summary": "大屏列表",
                "description": "",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "description": "身份令牌Token"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "required": false,
                        "type": "string",
                        "description": "大屏名称"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "total": {
                                            "type": "integer",
                                            "description": "总记录数",
                                            "format": "int32"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "登录令牌无效"
                    }
                }
            }
        },
        "/app/saiboard/admin/Screen/read": {
            "get": {
                "tags": [
                    "saiboard-admin.SAI Board.ScreenController"
                ],
                "summary": "大屏读取",
                "description": "",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "description": "身份令牌Token"
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "required": true,
                        "type": "integer",
                        "description": "大屏ID",
                        "format": "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "description": "大屏ID",
                                            "format": "int32"
                                        },
                                        "name": {
                                            "type": "string",
                                            "description": "大屏名称"
                                        },
                                        "draft_layout": {
                                            "type": "object",
                                            "description": "草稿布局"
                                        },
                                        "layout": {
                                            "type": "object",
                                            "description": "已发布布局"
                                        },
                                        "bg_config": {
                                            "type": "object",
                                            "description": "背景和适配配置"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "登录令牌无效"
                    }
                }
            }
        },
        "/app/saiboard/admin/Screen/save": {
            "post": {
                "tags": [
                    "saiboard-admin.SAI Board.ScreenController"
                ],
                "summary": "大屏添加",
                "description": "",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "description": "身份令牌Token"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "description": "",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "description": "大屏名称"
                                    },
                                    "bg_config": {
                                        "type": "object",
                                        "description": "背景配置：color/theme/fit_mode/fit_align/image/image_fit"
                                    }
                                },
                                "required": [
                                    "name"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": []
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "登录令牌无效"
                    }
                }
            }
        },
        "/app/saiboard/admin/Screen/update": {
            "put": {
                "tags": [
                    "saiboard-admin.SAI Board.ScreenController"
                ],
                "summary": "大屏修改",
                "description": "",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "description": "身份令牌Token"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "description": "",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "description": "大屏ID",
                                        "format": "int32"
                                    },
                                    "bg_config": {
                                        "type": "object",
                                        "description": "背景配置：color/theme/fit_mode/fit_align/image/image_fit"
                                    }
                                },
                                "required": [
                                    "id"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": []
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "登录令牌无效"
                    }
                }
            }
        },
        "/app/saiboard/admin/Screen/destroy": {
            "delete": {
                "tags": [
                    "saiboard-admin.SAI Board.ScreenController"
                ],
                "summary": "大屏删除",
                "description": "",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "description": "身份令牌Token"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "description": "",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "ids": {
                                        "type": "array",
                                        "description": "大屏ID列表",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "required": [
                                    "ids"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": []
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "登录令牌无效"
                    }
                }
            }
        },
        "/app/saiboard/admin/Screen/changeStatus": {
            "post": {
                "tags": [
                    "saiboard-admin.SAI Board.ScreenController"
                ],
                "summary": "大屏状态",
                "description": "",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "description": "身份令牌Token"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "description": "",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "description": "大屏ID",
                                        "format": "int32"
                                    },
                                    "status": {
                                        "type": "integer",
                                        "description": "状态，仅允许改为草稿2；发布请使用 publish",
                                        "format": "int32"
                                    }
                                },
                                "required": [
                                    "id",
                                    "status"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": []
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "登录令牌无效"
                    }
                }
            }
        },
        "/app/saiboard/admin/Screen/saveLayout": {
            "post": {
                "tags": [
                    "saiboard-admin.SAI Board.ScreenController"
                ],
                "summary": "保存大屏布局",
                "description": "",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "description": "身份令牌Token"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "description": "",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "description": "大屏ID",
                                        "format": "int32"
                                    },
                                    "layout": {
                                        "type": "object",
                                        "description": "布局JSON"
                                    }
                                },
                                "required": [
                                    "id",
                                    "layout"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": []
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "登录令牌无效"
                    }
                }
            }
        },
        "/app/saiboard/admin/Screen/publish": {
            "post": {
                "tags": [
                    "saiboard-admin.SAI Board.ScreenController"
                ],
                "summary": "发布大屏",
                "description": "",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "description": "身份令牌Token"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "description": "",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "description": "大屏ID",
                                        "format": "int32"
                                    }
                                },
                                "required": [
                                    "id"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": []
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "登录令牌无效"
                    }
                }
            }
        },
...
        "/app/saiboard/admin/Screen/versions": {
            "get": {
                "tags": [
                    "saiboard-admin.SAI Board.ScreenController"
                ],
                "summary": "大屏版本列表",
                "description": "",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "description": "身份令牌Token"
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "required": true,
                        "type": "integer",
                        "description": "大屏ID",
                        "format": "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": []
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "登录令牌无效"
                    }
                }
            }
        },
...
    },
...
}